+++ deflect/Server.h
@@ -1,5 +1,5 @@
 /*********************************************************************/
-/* Copyright (c) 2013-2015, EPFL/Blue Brain Project                  */
+/* Copyright (c) 2013-2016, EPFL/Blue Brain Project                  */
 /*                          Raphael Dumusc <raphael.dumusc@epfl.ch>  */
 /*                          Daniel.Nachbaur@epfl.ch                  */
 /* All rights reserved.                                              */
@@ -73,9 +73,6 @@ public:
 
     /** Destructor */
     ~Server();
-
-    /** Get the command handler. */
-    CommandHandler& getCommandHandler();
 
     /** Get the PixelStreamDispatcher. */
     FrameDispatcher& getPixelStreamDispatcher();