+++ doc/Changelog.md
@@ -1,13 +1,18 @@
 Changelog {#Changelog}
 ============
 
-## Deflect 0.10
+## Deflect 0.11
 
-### git master (0.10.2)
+### 0.11.0 (git master)
+* [94](https://github.com/BlueBrain/Deflect/pull/94):
+  Removed legacy SendCommand functionality from Stream to Server. It was only
+  used internally by the old Tide Dock streamer and never meant to be public.
 * [93](https://github.com/BlueBrain/Deflect/pull/93):
   Minor fixes for the Qml stream API:
   - Cleanup size handling, allow programmatic resizing of root Qml item
   - Do not send a size hints event if none were set by the user
+
+## Deflect 0.10
 
 ### 0.10.1 (01-04-2016)
 * [79](https://github.com/BlueBrain/Deflect/pull/79):